Version 0.2 - Score saving
1) Added score saving:

Now an actual score is displayed on the game over panel and the best score is saved.

2) Changed scoring mechanics:

Now the player gains score points when he passes over obstacle, not when he entries it.

3) Added camera animation:

When the game has been started after launching the application the main camera smoothly moves to the player but the game starts immediately after restarting.

Other changes:

4) Changed the design of the game over panel.

5) Now the player can't jump while the game on pause.